[수정] Gpt Temperature 값 변경

+++ src/main/java/com/YH/yeohaenghama/domain/chatAI/service/ChatAIInfo2.java
@@ -331,7 +331,7 @@ public class ChatAIInfo2 {
         plus + question,
         OpenAiChatOptions.builder()
             .withModel("gpt-3.5-turbo")
-            .withTemperature(0.4F)
+            .withTemperature((float) 0.4)
             .withMaxTokens(500)
             .build()
     );
